move util files into lib directory
