+++ client/src/Components/Events/EventPosterForm.tsx
@@ -3,8 +3,8 @@ import FormControlLabel from '@mui/material/FormControlLabel';
 import Switch from '@mui/material/Switch';
 import TextField from '@mui/material/TextField';
 import React, { useCallback, useEffect, useState } from 'react';
-import { usePostEvents } from '../../utils/apis/discord-api';
-import { useEventSettings } from '../../utils/apis/event-api';
+import { usePostEvents } from '../../lib/apis/discord-api';
+import { useEventSettings } from '../../lib/apis/event-api';
 import { ErrorMessage } from '../Common/ErrorMessage';
 import { useToast } from '../Common/ToastContext';
 import LoaderPage from '../LoaderPage';
