+++ client/src/Components/Roster/GuildMemberCard.tsx
@@ -10,12 +10,12 @@ import Typography from '@mui/material/Typography';
 import React, { useCallback, useState } from 'react';
 import { ReactComponent as DiscordLogo } from '../../assets/images/discord.svg';
 import gw2Image from '../../assets/images/gw2.png';
-import DiscordRole from '../../interfaces/discord-role';
-import MemberRecord from '../../interfaces/member-record';
-import { useUpdateDiscordMember } from '../../utils/apis/discord-api';
-import { useAddWarningMutation } from '../../utils/apis/warnings-api';
-import { getDateString } from '../../utils/data-processing';
-import { getColorFromRole } from '../../utils/helpers';
+import DiscordRole from '../../lib/interfaces/discord-role';
+import MemberRecord from '../../lib/interfaces/member-record';
+import { useUpdateDiscordMember } from '../../lib/apis/discord-api';
+import { useAddWarningMutation } from '../../lib/apis/warnings-api';
+import { getDateString } from '../../lib/utils/data-processing';
+import { getColorFromRole } from '../../lib/utils/helpers';
 import './GuildMemberCard.scss';
 import GuildMemberDetails from './GuildMemberDetails';
 import GuildMemberMenu from './GuildMemberMenu';
