+++ client/src/Components/Roster/Roster.tsx
@@ -2,9 +2,9 @@ import { useState } from 'react';
 import { useSearchParams } from 'react-router-dom';
 import AutoSizer from 'react-virtualized-auto-sizer';
 import { FixedSizeList as List } from 'react-window';
-import MemberRecord from '../../interfaces/member-record';
-import { useAuth } from '../../utils/apis/auth-api';
-import { useFilterString } from '../../utils/use-filter-string';
+import MemberRecord from '../../lib/interfaces/member-record';
+import { useAuth } from '../../lib/apis/auth-api';
+import { useFilterString } from '../../lib/utils/use-filter-string';
 import { ErrorMessage } from '../Common/ErrorMessage';
 import LoaderPage from '../LoaderPage';
 import GuildMemberCard from './GuildMemberCard';
