+++ client/src/lib/apis/event-api.ts
@@ -1,8 +1,8 @@
 import axios, { AxiosError } from 'axios';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
 import { useToast } from '../../Components/Common/ToastContext';
-import Event from '../../interfaces/event';
-import EventSettings from '../../interfaces/event-settings';
+import Event from '../interfaces/event';
+import EventSettings from '../interfaces/event-settings';
 
 export const useEvents = () => useQuery<Event[], AxiosError>('events');
 export const useEventSettings = () => useQuery<EventSettings, AxiosError>('events/settings');
