+++ client/src/lib/apis/gw2-api.ts
@@ -1,8 +1,8 @@
 import { AxiosError } from 'axios';
 import { useQuery } from 'react-query';
-import GW2LogEntry from '../../interfaces/gw2-log-entry';
-import GW2Member from '../../interfaces/gw2-member';
-import GW2Rank from '../../interfaces/gw2-rank';
+import GW2LogEntry from '../interfaces/gw2-log-entry';
+import GW2Member from '../interfaces/gw2-member';
+import GW2Rank from '../interfaces/gw2-rank';
 
 export const useGW2Members = () => useQuery<GW2Member[], AxiosError>('gw2/members');
 export const useGW2Log = () => useQuery<GW2LogEntry[], AxiosError>('gw2/log');
