+++ client/src/lib/interfaces/auth-info.ts
