+++ client/src/lib/interfaces/discord-log-string-generator.ts
