+++ client/src/lib/interfaces/discord-log.ts
