+++ client/src/lib/interfaces/discord-member.ts
