+++ client/src/lib/interfaces/discord-role.ts
