+++ client/src/lib/interfaces/event-settings.ts
