+++ client/src/lib/interfaces/event.ts
