+++ client/src/lib/interfaces/gw2-log-entry.ts
