+++ client/src/lib/interfaces/gw2-member.ts
