+++ client/src/lib/interfaces/gw2-rank.ts
