+++ client/src/lib/interfaces/member-info.ts
