+++ client/src/lib/interfaces/member-record.ts
