+++ client/src/lib/interfaces/recruitment-post.ts
