+++ client/src/lib/utils/discord-log-string-factory.ts
