+++ client/src/lib/utils/helpers.ts
